changed admin files to require processadmin
