+++ pages/admin/registerStudentFinancial.php
@@ -26,7 +26,7 @@
 <br>
 
 <form method="post" action="" class ="col s8">
-  <?php require_once('../../unsecure/processunsecure.php'); ?>
+  <?php require_once('processadmin.php');?>
 
   <div class="row">
    <div class="input-field col s4">
