+++ pages/admin/viewInventory.php
@@ -37,5 +37,5 @@
 </form>
 
 <?php
-require_once('../../unsecure/processunsecure.php');
+require_once('processadmin.php');
 ?>
