+++ pages/admin/viewStudentPersonal.php
@@ -35,9 +35,6 @@
 </div>
 </form>
 
-<?php
-require_once('../../unsecure/processunsecure.php');
-?>
-
+<?php require_once('processadmin.php');?>
 </body>
 </html>
